pipeline/requirements.txt
pandas
pyarrow
geopandas
shapely
requests
numpy
loguru
google-cloud-bigquery
dbt-core
dbt-bigquery
